requirements.txt
click
colorama
edge-tts
groq
langchain-community
pyaudio
pydub
pygame
python-dotenv
speechrecognition
